Write a function that returns the length of a string.

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -6,9 +6,9 @@
  * Return: The length of @str.
  */
 
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
